Add delete operations
